onBeforeCopy moved upper BorderedBox component
